# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.ArizaDetay:
    properties:
      Aciklama:
        type: string
      ArizaDurumu:
        type: string
      ArizaID:
        type: integer
      ArizaTipi:
        type: string
      ArizaTipiTanimAdi:
        type: string
      ArizaTipiTanimID:
        type: integer
      BlokTarafi:
        type: string
      Bloklar:
        items:
          type: integer
        type: array
      CreateUserName:
        type: string
      Detay: {}
      IsEmriOlusturmaTarihi:
        type: string
      MakineAdi:
        type: string
      MakineID:
        type: integer
      Pozisyonlar:
        items:
          type: integer
        type: array
    type: object
...
  models.GetArizaRequestBody:
    properties:
      ArizaID:
        type: integer
      ArizaStatusu:
        type: string
      UserID:
        type: integer
    type: object
...
host: localhost:45013
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a microservice.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sasa El Terminali API
  version: "1.0"
paths:
...
  /api/v1/ariza/get:
    post:
      consumes:
      - application/json
      description: Ariza getir
      parameters:
      - description: ArizaID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.GetArizaRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ArizaDetay'
      security:
      - Bearer: []
      summary: Get ariza by ID
      tags:
      - ariza
...
securityDefinitions:
  Bearer:
    description: '"Type ''Bearer'' followed by a space and then your token."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"